% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_las_data.R
\name{check_las_data}
\alias{check_las_data}
\title{Check a for a directory with .laz|.las files,
the path of a file, or an object of class \code{LAS}|\code{LASCatalog}}
\usage{
check_las_data(las)
}
\arguments{
\item{las}{character. a directory with .laz|.las files files,
the path of a single .laz|.las file, -or- an object of class \code{LAS}|\code{LASCatalog}}
}
\description{
Check a character for a directory with .laz|.las files,
the path of a file, or an object of class \code{LAS}|\code{LASCatalog}
}
\keyword{internal}
