% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spatial_points.R
\name{check_spatial_points}
\alias{check_spatial_points}
\title{Check a data frame for spatial point data. Convert to points if needed.}
\usage{
check_spatial_points(tree_list, crs = NA)
}
\arguments{
\item{tree_list}{data.frame. A data frame with the columns \code{treeID}, \code{tree_x}, \code{tree_y}, and \code{tree_height_m}.
If an \code{sf} class object with POINT geometry (see \code{\link[sf:st_geometry_type]{sf::st_geometry_type()}}), the program will use the data "as-is" and only require the \code{treeID} and \code{tree_height_m} columns.}

\item{crs}{string. A crs string as returned from \code{\link[sf:st_crs]{sf::st_crs()}} or the EPSG code of the x,y coordinates.
Defaults to the crs of the \code{tree_list} data if of class "sf".}
}
\description{
Check a data frame for spatial point data. Convert to points if needed.
}
\keyword{internal}
