% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_las_catalog.R
\name{chunk_las_catalog}
\alias{chunk_las_catalog}
\title{Tile raw \code{.las}|\code{.laz} files to work with smaller chunks}
\usage{
chunk_las_catalog(
  folder,
  outfolder = getwd(),
  accuracy_level = 2,
  max_ctg_pts = 7e+07,
  max_area_m2 = 9e+07,
  transform = FALSE,
  new_crs = NA,
  old_crs = NA
)
}
\arguments{
\item{folder}{string. The path of a folder containing a set of las/laz files. Can also be a vector of file paths.}

\item{outfolder}{string. The path of a folder to write the tiled las files to.}

\item{accuracy_level}{numeric. Choose processing accuracy.
accuracy_level = 1 uses DTM to height normalize the points
accuracy_level = 2 uses triangulation with high point density (20 pts/m2) to height normalize the points
accuracy_level = 3 uses triangulation with very high point density (100 pts/m2) to height normalize the points}

\item{max_ctg_pts}{numeric. Max number of points to process at one time. Setting this number higher will possibly reduce run times but increase the chance of running out of memory and vice versa.}

\item{max_area_m2}{numeric. Max area to process at one time. See \code{max_ctg_pts} parameter, this one is less important as never experienced memory issues with large areas (just lots of points)}

\item{transform}{logical. should the las/laz files be transformed? If set to \code{TRUE} the parameters \code{new_crs} must be defined.}

\item{new_crs}{string. crs to change to as an epsg numerical code}

\item{old_crs}{string. crs to change from as an epsg numerical code}
}
\value{
A list of 1) \code{process_data} an sf object; 2) \code{is_chunked_grid} indicator if chunks were created; 3) \code{plt} a ggplot object
}
\description{
Function to tile raw \code{.las}|\code{.laz} files to work with smaller chunks based on point density and coverage area
}
\examples{
 \dontrun{
 f <- "../lasdata"
 chunk_las_catalog(folder = f, outfolder = getwd())
 }
}
\references{
\url{https://r-lidar.github.io/lidRbook/norm.html}
\url{https://github.com/r-lidar/lasR/issues/18#issuecomment-2027818414}
}
