% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aoi.R
\name{clip_tree_list_aoi}
\alias{clip_tree_list_aoi}
\title{internal functions to work with tree list data within a specified area of interest (AOI) or "domain" in the terminology of really smart fire modellers.}
\usage{
clip_tree_list_aoi(
  tree_list,
  crs,
  study_boundary,
  bbox_aoi = F,
  buffer = 0,
  reproject_epsg = NULL
)
}
\arguments{
\item{tree_list}{data.frame. A data frame with the columns \code{treeID}, \code{tree_x}, and \code{tree_y}.
If an \code{sf} class object with POINT geometry (see \code{\link[sf:st_geometry_type]{sf::st_geometry_type()}}), the program will use the data "as-is" and only require the \code{treeID} column.
Other required columns include:
\itemize{
\item \code{crown_area_m2}, \code{tree_height_m} (e.g. as exported by \code{\link[=raster2trees]{raster2trees()}})
\item \code{tree_cbh_m} (e.g. as exported by \code{\link[=trees_cbh]{trees_cbh()}})
\item and one of \code{dbh_cm}, \code{dbh_m}, or  \code{basal_area_m2} (e.g. as exported by \code{\link[=trees_dbh]{trees_dbh()}})
}}

\item{crs}{string. A crs string as returned from \code{\link[sf:st_crs]{sf::st_crs()}} or the EPSG code of the x,y coordinates.
Defaults to the crs of the \code{tree_list} data if of class "sf".}

\item{study_boundary}{sf. The boundary of the study area to define the area of interest which may extend beyond the space with trees.
This must be an sf class object with a single record. If you need to get the trees within multiple different AOI's, then \code{purrr::map()} this function.}

\item{bbox_aoi}{logical. Should the study_boundary be transformed to a bounding box instead of it's original shape for determining the trees within the boundary?
If set to true, the bounding box is created prior to applying the buffer.}

\item{buffer}{numeric. Buffer to be applied to the study area prior to determining trees within the boundary.
Units are determined by the horizontal CRS settings of the tree_list data or the CRS of the reproject_epsg.}

\item{reproject_epsg}{numeric. The EPSG code to reproject the data in prior to buffering and clipping.
Will determine the projection of the output data.}
}
\description{
internal functions to work with tree list data within an AOI.
For example, to use the tree list (e.g. as exported by \code{\link[=raster2trees]{raster2trees()}}) within the QUIC-Fire modelling tool some additional columns are required
and one needs to define a study area and align the DTM and tree list with this domain.
}
\keyword{internal}
