% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_structure.R
\name{create_project_structure}
\alias{create_project_structure}
\title{Create project structure}
\usage{
create_project_structure(
  output_dir,
  input_las_dir,
  input_treemap_dir = paste0(system.file(package = "cloud2trees"), "/extdata/treemap")
)
}
\arguments{
\item{output_dir}{parent directory where new folders \code{point_cloud_processing_delivery} and \code{point_cloud_processing_temp} will be written for exports}

\item{input_las_dir}{directory where .las|.laz point cloud data exists...program will search all sub-directories for all .las|.laz files and process them as one}

\item{input_treemap_dir}{directory where Treemap 2016 exists. Use \code{\link[=get_treemap]{get_treemap()}} first.}
}
\value{
A data.frame.
}
\description{
Function to generate nested project directories based on the user-defined directory to create output file structure
}
\keyword{internal}
