% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_biomass.R
\name{get_cruz_stand_kg_per_m3}
\alias{get_cruz_stand_kg_per_m3}
\title{internal functions to estimate tree biomass}
\usage{
get_cruz_stand_kg_per_m3(
  forest_type_group_code,
  basal_area_m2_per_ha,
  trees_per_ha
)
}
\arguments{
\item{forest_type_group_code}{numeric. as extracted by \code{\link[=trees_type]{trees_type()}}}

\item{basal_area_m2_per_ha}{numeric.}

\item{trees_per_ha}{numeric.}
}
\description{
internal functions to extract raster values at point locations used by \verb{trees_biomass*()} functions
for the most part, these functions are used to distribute raster cell (i.e. "stand") level estimates of fuel load
to individual trees
}
\keyword{internal}
