% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download all external data used by package}
\usage{
get_data(savedir = NULL, force = F)
}
\arguments{
\item{savedir}{Optional directory to save data in a new location. Defaults to package contents.}

\item{force}{Whether to overwrite existing data}
}
\description{
The package requires external data to estimate individual tree DBH (see \code{\link[=get_treemap]{get_treemap()}}) and to extract the forest type for a tree list (see \code{\link[=get_foresttype]{get_foresttype()}}).
This is a all-in-one function that downloads all of the external data used by the package.
}
\examples{
 \dontrun{
 get_data()
 }
}
