% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landfire.R
\name{get_landfire}
\alias{get_landfire}
\title{Download Forest Type Groups of the Continental United States data}
\usage{
get_landfire(savedir = NULL, force = F)
}
\arguments{
\item{savedir}{Optional directory to save data in a new location. Defaults to package contents.}

\item{force}{Whether to overwrite existing data}
}
\description{
The LANDFIRE Forest Canopy Bulk Density (CBD) data is used to estimate individual tree crown biomass in kilograms
See \code{\link[=trees_biomass_landfire]{trees_biomass_landfire()}}
}
\examples{
 \dontrun{
 get_landfire()
 }
}
\references{
\itemize{
\item \href{https://landfire.gov/fuel/cbd}{LANDFIRE Forest Canopy Bulk Density (CBD)}
U.S. Department of Agriculture and U.S. Department of the Interior.
}
}
