% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_data.R
\name{get_url_data}
\alias{get_url_data}
\title{Download url data}
\usage{
get_url_data(
  eval_url = NULL,
  my_name = NULL,
  savedir = NULL,
  req_file_list = NULL,
  force = F,
  cleanup_zip = T,
  move_files_to_top = T
)
}
\arguments{
\item{eval_url}{Required url of the .zip file to be downloaded. url string must end in .zip}

\item{my_name}{Required name for the folder which data will be extracted to}

\item{savedir}{Optional directory to save data in a new location. Defaults to package contents.}

\item{req_file_list}{Optional list of files to check for before re-downloading full data}

\item{force}{Whether to overwrite exising data}

\item{cleanup_zip}{Whether to remove the .zip file after extracting the contents}
}
\description{
Generic function to download data from a url with .zip file data
Functions \code{\link[=get_treemap]{get_treemap()}} and \code{\link[=get_foresttype]{get_foresttype()}} use this
}
\examples{
 \dontrun{
 get_treemap()
 }

}
\keyword{internal}
