% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_projection.R
\name{get_vertical_crs}
\alias{get_vertical_crs}
\title{tools for working with projections, crs, epsg, and whatnot}
\usage{
get_vertical_crs(x)
}
\arguments{
\item{x}{sf. or any object of a class that can be read by sf::st_crs such as "LAS" class}
}
\value{
A data.frame with the units and scale factor for the vertical projection
}
\description{
tools for working with projections, crs, epsg, and whatnot
}
\keyword{internal}
