% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladderfuelsr_cbh.R
\name{ladderfuelsr_cbh}
\alias{ladderfuelsr_cbh}
\title{estimate CBH using \code{LadderFuelsR} package}
\usage{
ladderfuelsr_cbh(
  las,
  treeID = NA,
  min_vhp_n = 4,
  voxel_grain_size_m = 2,
  dist_btwn_bins_m = 1,
  min_fuel_layer_ht_m = 1,
  lad_pct_gap = 25,
  lad_pct_base = 25,
  num_jump_steps = 1,
  min_lad_pct = 10,
  frst_layer_min_ht_m = 1
)
}
\arguments{
\item{las}{string -or- object. a .las|.laz file path -OR- an object of class LAS that has been height normalized.}

\item{treeID}{string. the LadderFuelsR process requires a treeID that uniquely identifies points within a tree, if left as NA this process will attempt to locate the \code{treeID} data based on an attribute in the point cloud.}

\item{min_vhp_n}{numeric. the minimum number of vertical height profiles (VHPs) needed to estimate a CBH.}

\item{voxel_grain_size_m}{numeric. horizontal resolution (suggested 1 meter for lad profiles and 10 meters for LAI maps). See \code{grain.size} in \code{\link[leafR:lad.voxels]{leafR::lad.voxels()}}}

\item{dist_btwn_bins_m}{numeric. value for the actual height bin step (in meters). See \code{step} in \code{\link[LadderFuelsR:get_gaps_fbhs]{LadderFuelsR::get_gaps_fbhs()}}}

\item{min_fuel_layer_ht_m}{numeric. value for the actual minimum base height (in meters). See \code{min_height} in \code{\link[LadderFuelsR:get_gaps_fbhs]{LadderFuelsR::get_gaps_fbhs()}}}

\item{lad_pct_gap}{numeric. value of the percentile threshold used to identify gaps (default percentile 25th). See \code{perc_gap} in \code{\link[LadderFuelsR:get_gaps_fbhs]{LadderFuelsR::get_gaps_fbhs()}}}

\item{lad_pct_base}{numeric. value of the percentile threshold used to identify fuels layers base height (default percentile 25th). See \code{perc_base} in \code{\link[LadderFuelsR:get_gaps_fbhs]{LadderFuelsR::get_gaps_fbhs()}}}

\item{num_jump_steps}{numeric. value for the number of height bin steps that can be jumped to reshape fuels layers. See \code{number_steps} in \code{\link[LadderFuelsR:get_real_fbh]{LadderFuelsR::get_real_fbh()}}}

\item{min_lad_pct}{numeric. value for the minimum required LAD percentage in a fuel layer. See \code{threshold} in \code{\link[LadderFuelsR:get_layers_lad]{LadderFuelsR::get_layers_lad()}}}

\item{frst_layer_min_ht_m}{numeric. value for the depth height of the first fuel layer.
If the first fuel layer has the maximum LAD and its depth is greater than the indicated value
, then this fuel layer is considered as the CBH of the tree. On the contrary, if its depth is <= the value, the CBH with maximum LAD will be the second fuel layer, although it has not the maximum LAD.
See \code{hdepth1_height} in \code{\link[LadderFuelsR:get_cbh_metrics]{LadderFuelsR::get_cbh_metrics()}}}
}
\value{
Returns an a list of data.frame objects that are the results of the different LadderFuelsR steps. Returns NULL if the process is unable to detect a CBH from the point cloud.
}
\description{
\code{ladderfuelsr_cbh()} is an all-in-one function to process height normalized .las|.laz files
using the functionality of the \code{LadderFuelsR} package.
The function returns a a list of data.frame objects that are the results of the different LadderFuelsR steps. Returns NULL if the process is unable to detect a CBH from the point cloud.
\code{ladderfuelsr_cbh()} outputs:

The order of operations is:
\itemize{
\item Create a data frame of the 3D voxels information (xyz) with Leaf Area Density (LAD) values from las file using \code{\link[leafR:lad.voxels]{leafR::lad.voxels()}}
\item Calculate the lad profile from the input lad.voxels using \code{\link[leafR:lad.profile]{leafR::lad.profile()}}
\item Calculate gaps and fuel layers base height (FBH) as the difference in percentiles between consecutive LAD values along the vertical tree profile (VTP) using \code{\link[LadderFuelsR:get_gaps_fbhs]{LadderFuelsR::get_gaps_fbhs()}}
\item Calculate the percentile value of each fuel layer base height using \code{\link[LadderFuelsR:calculate_gaps_perc]{LadderFuelsR::calculate_gaps_perc()}}
\item Calculate distances (and their heights) between fuel layers as the difference between consecutive gaps and fuel bases using \code{\link[LadderFuelsR:get_distance]{LadderFuelsR::get_distance()}}
\item Calculate fuels depth as the difference between gaps interleaved between fuel layers minus one step if the fuel depths are greater than one step using \code{\link[LadderFuelsR:get_depths]{LadderFuelsR::get_depths()}}
\item Reshape fuel layers after removing distances equal to any number of height bin steps, keeping the first "base height" from those consecutive ones separated by such distance using \code{\link[LadderFuelsR:get_real_fbh]{LadderFuelsR::get_real_fbh()}}
\item Recalculate fuel layers depth after considering distances greater than the actual height bin step using \code{\link[LadderFuelsR:get_real_depths]{LadderFuelsR::get_real_depths()}}
\item Recalculate the distance between fuel layers after considering distances greater than any number of height bin steps using \code{\link[LadderFuelsR:get_effective_gap]{LadderFuelsR::get_effective_gap()}}
\item Calculate the percentage of LAD within each fuel layer (first output) and removes those fuel layers with LAD percentage less than a specified threshold using \code{\link[LadderFuelsR:get_layers_lad]{LadderFuelsR::get_layers_lad()}}
\item Determine the CBH of a segmented tree using three criteria: maximum LAD percentage, maximum distance and the last distance using \code{\link[LadderFuelsR:get_cbh_metrics]{LadderFuelsR::get_cbh_metrics()}}
}
}
\references{
\itemize{
\item \url{https://doi.org/10.1111/2041-210X.14427}
Viedma, O., Silva, C. A., Moreno, J. M., & Hudak, A. T. (2024). LadderFuelsR: A new automated tool for vertical fuel continuity analysis and crown base height detection using light detection and ranging. Methods in Ecology and Evolution.
\url{https://github.com/olgaviedma/LadderFuelsR}
\item \url{https://doi.org/10.3390/rs11010092}
Almeida, D. R. A. D., Stark, S. C., Shao, G., Schietti, J., Nelson, B. W., Silva, C. A., ... & Brancalion, P. H. S. (2019). Optimizing the remote detection of tropical rainforest structure with airborne lidar: Leaf area profile sensitivity to pulse density and spatial sampling. Remote Sensing, 11(1), 92.
\url{https://github.com/DRAAlmeida/leafR}
}
}
