% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasr_dtm_norm.R
\name{lasr_dtm_norm}
\alias{lasr_dtm_norm}
\title{use lasR to combine DTM and normalize step}
\usage{
lasr_dtm_norm(dtm_file_name, frac_for_tri = 1, dtm_res = 1, norm_accuracy = 2)
}
\arguments{
\item{dtm_file_name}{string. Where to write the DTM.}

\item{frac_for_tri}{numeric. The fraction of points used in Delauny triangulation.}

\item{dtm_res}{numeric. The desired resolution of the DTM produced in meters.}

\item{norm_accuracy}{numeric. see \code{chunk_las_catalog}. Choose processing accuracy.
accuracy_level = 1 uses DTM to height normalize the points
accuracy_level = 2 uses triangulation with high point density (20 pts/m2) to height normalize the points
accuracy_level = 3 uses triangulation with very high point density (100 pts/m2) to height normalize the points}
}
\value{
A \code{lasR} pipeline
}
\description{
Combining DTM and normalize step using \code{lasR} functionality
}
\references{
https://r-lidar.github.io/lidRbook/norm.html
https://github.com/r-lidar/lasR/issues/18#issuecomment-2027818414
}
\keyword{internal}
