% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasr_pipeline.R
\name{lasr_pipeline}
\alias{lasr_pipeline}
\title{Create \code{lasR} package pipeline to process a las grid tile created via \code{\link[=chunk_las_catalog]{chunk_las_catalog()}}}
\usage{
lasr_pipeline(
  processing_grid_num = 1,
  process_data,
  keep_intrmdt = F,
  dtm_res_m = 1,
  chm_res_m = 0.25,
  min_height = 2,
  max_height = 70,
  noise_level = 2,
  dtm_dir = getwd(),
  chm_dir = getwd(),
  classify_dir = getwd(),
  normalize_dir = getwd()
)
}
\arguments{
\item{processing_grid_num}{numeric. processing_grid column in the data.frame created via \code{\link[=chunk_las_catalog]{chunk_las_catalog()}}}

\item{process_data}{data.frame. data.frame created via \code{\link[=chunk_las_catalog]{chunk_las_catalog()}}}

\item{keep_intrmdt}{logical. this process writes intermediate data to the disk, keep those intermediate files (classfied, normalized, stem las files)?}

\item{dtm_res_m}{numeric. The desired resolution of the DTM produced in meters.}

\item{chm_res_m}{numeric. The desired resolution of the CHM produced in meters.}

\item{min_height}{numeric. Set the minimum height (m) for individual tree detection}

\item{max_height}{numeric. Set the maximum height (m) for the canopy height model}

\item{noise_level}{numeric. Choose point cloud noise reduction level 1, 2, or 3. Use a higher noise level for point clouds with more noise
which tend to produce raster outputs with pits or spikes that are too severe to be filled with standard post-processing.
The default level of 2 has similar processing time compared to level to 1
but uses a more broadly applicable noise detection algorithm. Level 3 takes 10-40\% longer to process.
\itemize{
\item noise_level = 1 uses isolated voxel filter (IVF) with a resolution of 5 voxels and 9 other points to identify noise
\item noise_level = 2 uses a single-pass, fine-scale statistical outlier removal (SOR) that primarily targets local noise
\item noise_level = 3 uses a muli-pass statistical outlier removal (SOR) that first applies a coarse-scale filter
to find points/clusters far from the main cloud mass and then applies a fine-scale filter to identify local noise
}}

\item{dtm_dir}{string. The path of a folder to write the tiled DTM files to.}

\item{chm_dir}{string. The path of a folder to write the tiled CHM files to.}

\item{classify_dir}{string. The path of a folder to write the classified .las files to.}

\item{normalize_dir}{string. The path of a folder to write the normalized .las files to.}
}
\value{
A \code{lasR} pipeline answer list
}
\description{
Create \code{lasR} package pipeline to process a las grid tile created via \code{\link[=chunk_las_catalog]{chunk_las_catalog()}}
}
\references{
\url{https://r-lidar.github.io/lasR/index.html}
}
\keyword{internal}
