% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafr_for_ladderfuelsr.R
\name{leafr_for_ladderfuelsr}
\alias{leafr_for_ladderfuelsr}
\title{re-writes \code{leafR} steps to allow for treeID as input for \code{ladderfuelsR} or \code{\link[=ladderfuelsr_cbh]{ladderfuelsr_cbh()}}}
\usage{
leafr_for_ladderfuelsr(
  las,
  voxel_grain_size_m = 1,
  k = 1,
  attribute = "treeID",
  min_pulses = 0,
  relative = FALSE
)
}
\arguments{
\item{las}{an object of class LAS that has been height normalized.}

\item{voxel_grain_size_m}{numeric. horizontal resolution (suggested 1 meter for lad profiles and 10 meters for LAI maps). See \code{grain.size} in \code{\link[leafR:lad.voxels]{leafR::lad.voxels()}}}

\item{k}{numeric. coefficient to transform effective LAI to real LAI (k = 1; for effective LAI)}

\item{attribute}{character. The column name of the attribute to group the return LAD profile by. Default is "treeID". The attribute (whatever it is defined as)
must exist in the \code{las} data}

\item{min_pulses}{numeric. minimum number of pulses required to return a record by attribute. set to zero (default) to leave data unfiltered.}

\item{relative}{logical. produce lad profile by relative total LAI values. Indicate when using effective LAI value.
if set to TRUE, lad value will be relative_lad; otherwise, lad value will be mean_lad}
}
\value{
Returns an data.frame which can have multiple treeIDs to use as input for \code{\link[=ladderfuelsr_cbh]{ladderfuelsr_cbh()}}
}
\description{
\code{leafr_for_ladderfuelsr()} is a re-write of \code{\link[leafR:lad.voxels]{leafR::lad.voxels()}} and \code{\link[leafR:lad.profile]{leafR::lad.profile()}} to:
\itemize{
\item removes the requirement to use a file written to disk
\item allows for the calculation of LAD by the \code{treeID} attribute so that don't have to pass individual tree point clouds
\item updates to the use of the latest \code{lidR} functionality and removes the use of \code{sp} and \code{raster} functions
\item updates the function to \code{tidy} data manipulation
}
}
\examples{
 \dontrun{
 # polygon data
 f <- system.file(package = "cloud2trees","extdata","crowns_poly.gpkg")
 trees_poly <- sf::st_read(f)
 # simplify polygons
 trees_poly <- simplify_multipolygon_crowns(trees_poly)
 # point cloud data
 lf <- system.file(package = "cloud2trees","extdata","norm_las","RMNP_017_2018_normalize.las")
 las <- lidR::readLAS(lf)
 las@data \%>\% dplyr::glimpse()
 # polygon_attribute_to_las to attach treeID to las
 las <- polygon_attribute_to_las(las, trees_poly, force_crs = T, attribute = "treeID")
 las@data \%>\% dplyr::glimpse()
 # get the lad profile for each treeID
 lad_profile <- leafr_for_ladderfuelsr(
     las
     , voxel_grain_size_m = 1
     , k = 1
     , group_treeID = T
     , relative = F
   )
 dplyr::glimpse(lad_profile)
 }
}
\references{
\itemize{
\item \url{https://doi.org/10.3390/rs11010092}
Almeida, D. R. A. D., Stark, S. C., Shao, G., Schietti, J., Nelson, B. W., Silva, C. A., ... & Brancalion, P. H. S. (2019). Optimizing the remote detection of tropical rainforest structure with airborne lidar: Leaf area profile sensitivity to pulse density and spatial sampling. Remote Sensing, 11(1), 92.
\url{https://github.com/DRAAlmeida/leafR}
}
}
