% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piles_spectral_filter.R
\name{piles_spectral_filter}
\alias{piles_spectral_filter}
\title{Spectral filtering of candidate slash piles
Spectral refinement of candidate slash piles}
\usage{
piles_spectral_filter(
  sf_data,
  rgb_rast,
  red_band_idx,
  green_band_idx,
  blue_band_idx,
  spectral_weight = 4,
  filter_return = T
)
}
\arguments{
\item{sf_data}{An sf object of candidate polygons to be filtered.}

\item{rgb_rast}{A multi-band SpatRaster containing RGB imagery.}

\item{red_band_idx}{Integer. Index of the red band.}

\item{green_band_idx}{Integer. Index of the green band.}

\item{blue_band_idx}{Integer. Index of the blue band.}

\item{spectral_weight}{Integer (0 to 6). The consensus threshold
requiring a specific number of index thresholds to be met for a candidate
pile to be retained.}

\item{filter_return}{Logical. If TRUE (default), the function returns
only polygons where the "inrange_th_votes" is greater than or equal to
the \code{spectral_weight}. If FALSE, the function returns the entire input
dataset with columns for each calculated spectral index and the
"inrange_th_votes" column for post-processing and analysis.}
}
\value{
An sf object. Depending on \code{filter_return}, this will either be
a subset of the input data or the full input data with appended
spectral voting columns.
}
\description{
Evaluates candidate segments generated by \code{\link[=piles_detect]{piles_detect()}}
against an RGB raster. It uses a voting system based on six spectral indices
to separate woody slash material from background vegetation or ground. This
function can either return a filtered dataset or the full input dataset
with appended spectral metrics for diagnostic purposes.
}
