% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2trees.R
\name{raster2trees}
\alias{raster2trees}
\title{Use a CHM raster to detect individual trees}
\usage{
raster2trees(
  chm_rast,
  outfolder,
  ws = itd_ws_functions()[["exp_fn"]],
  min_height = 2,
  min_crown_area = 0.1,
  tempdir = tempdir()
)
}
\arguments{
\item{chm_rast}{raster. A  raster from \code{terra} or \code{stars}representing a canopy height model}

\item{outfolder}{string. The path of a folder to write the crown vector data to}

\item{ws}{numeric or function. Length or diameter of the moving window used to detect the local
maxima in the units of the input data (usually meters). If it is numeric a fixed window size is used.
If it is a function, the function determines the size of the window at any given location on the canopy.
By default function takes the height of a given pixel as its only argument and return the
desired size of the search window when centered on that pixel.}

\item{min_height}{numeric. Set the minimum height (m) for individual tree detection}

\item{min_crown_area}{numeric. Set the minimum crown area (m2) for individual tree detection}

\item{tempdir}{string. Directory to write intermediate files. Intermediate files are only created for large rasters too big to fit in memory.}
}
\value{
Returns a spatial data frame of individual tree crown vectors detected using the CHM.
The tree top point coordinates are located in the \code{tree_x} and \code{tree_y} columns.
The process also writes two \code{.gpkg} files to the \code{outfolder} directory: \code{chm_detected_crowns.gpkg} and \code{chm_detected_tree_tops.gpkg}
}
\description{
\code{raster2trees()} is an all-in-one function to process a CHM raster
and return a spatial data frame of tree crown polygons and points.
The order of operations is:
\itemize{
\item Perform individual tree detection using \code{\link[lidR:locate_trees]{lidR::locate_trees()}} with the \code{\link[lidR:itd_lmf]{lidR::lmf()}} algorithm
\item Delineate tree crowns using \code{\link[ForestTools:mcws]{ForestTools::mcws()}}
}

Note, this function does not estimate DBH for the detected trees and only returns tree location, crown area, and height information.
To estimate tree DBH from the detected tree heights see \code{\link[=trees_dbh]{trees_dbh()}}.
}
\examples{
 \dontrun{
 f <- paste0(system.file(package = "cloud2trees"),"/extdata/chm.tif")
 crowns_sf <- raster2trees(chm_rast = terra::rast(f), outfolder = tempdir())
 crowns_sf \%>\% class()
 crowns_sf \%>\% dplyr::glimpse()
 crowns_sf \%>\% ggplot2::ggplot() + ggplot2::geom_sf(ggplot2::aes(fill=tree_height_m))
 }
}
\references{
\url{https://r-lidar.github.io/lidRbook/itd.html}
}
