% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproject_las.R
\name{reproject_las}
\alias{reproject_las}
\title{function to reproject las}
\usage{
reproject_las(filepath, new_crs = NA, old_crs = NA, outdir = getwd())
}
\arguments{
\item{filepath}{the full file path of the .las|.laz file}

\item{new_crs}{crs to change to as an epsg numerical code}

\item{old_crs}{crs to change from as an epsg numerical code}

\item{outdir}{where to save this re-projected .las file}
}
\value{
A character
}
\description{
Function to reproject las data using \code{lidR} but be careful! This is inefficient and potentially causes inaccuracies due to transformations (see reference).
}
\references{
\url{https://gis.stackexchange.com/questions/371566/can-i-re-project-an-las-file-in-lidr}
}
\keyword{internal}
