% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_multipolygon_crowns.R
\name{simplify_multipolygon_crowns}
\alias{simplify_multipolygon_crowns}
\title{Simplify MULTIPOLYGON to POLYGON geometry in an \code{sf} class object}
\usage{
simplify_multipolygon_crowns(trees_poly)
}
\arguments{
\item{trees_poly}{data.frame. A data frame of \code{sf} class with POLYGON,MULTIPOLYGON geometry (see \code{\link[sf:st_geometry_type]{sf::st_geometry_type()}}) and the column \code{treeID}}
}
\value{
A \code{sf} class object data frame
}
\description{
Function to simplify MULTIPOLYGON geometry to POLYGON geometry in an \code{sf} class object by selecting the largest segment of the MULTIPOLYGON
}
\examples{
 \dontrun{
 f <- paste0(system.file(package = "cloud2trees"),"/extdata/crowns_poly.gpkg")
 crowns <- sf::st_read(f, quiet = T)
 crowns \%>\% sf::st_geometry_type() \%>\% table()
 crowns_simp <- simplify_multipolygon_crowns(crowns)
 crowns_simp \%>\% sf::st_geometry_type() \%>\% table()
 }
}
