% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treels_stem_dbh.R
\name{treels_stem_dbh}
\alias{treels_stem_dbh}
\title{detect tree stems and estimate DBH using \code{TreeLS} package}
\usage{
treels_stem_dbh(
  folder,
  outfolder,
  min_height = 2,
  max_dbh = 2,
  chunk_these = FALSE
)
}
\arguments{
\item{folder}{string. The path of a folder containing a set of las/laz files. Can also be a vector of file paths.}

\item{outfolder}{string. The path of a folder to write the tiled vector files to}

\item{min_height}{numeric. Set the minimum height (m) for individual tree detection}

\item{max_dbh}{numeric. Set the largest tree diameter (m) expected in the point cloud}

\item{chunk_these}{logical. Do the las/laz files need to be tiled to work with smaller subsets? See \code{is_chunked_grid} in \code{\link[=chunk_las_catalog]{chunk_las_catalog()}}}
}
\value{
Returns an \code{sf} data.frame with TreeLS detected trees and DBH estimated directly from the point cloud.
Exports files in the
\code{outfolder} defined by the user in the function call.
}
\description{
\code{treels_stem_dbh()} is an all-in-one function to process height normalized .las|.laz files
using the functionality of the \code{TreeLS} package. The function generates a list of stems and estimates DBH directly from the point cloud.
\code{treels_stem_dbh()} outputs:
\itemize{
\item A .laz file with the \code{Classification} data updated to: ground points (class 2); water points (class 9); stem points (class 4); non-stem (class 5).
\item A vector data file in \code{gpkg} format with the tree identification stem locations, heights, and DBH estimates.
}

The order of operations is:
\itemize{
\item Detect tree stems/boles from the height normalized point cloud using \code{\link[TreeLS:treeMap]{TreeLS::treeMap()}} with the \code{\link[TreeLS:map.hough]{TreeLS::map.hough()}}  algorithm
\item Merge overlapping tree coordinates using \code{\link[TreeLS:treeMap.merge]{TreeLS::treeMap.merge()}}
\item Assign tree IDs to the original points using \code{\link[TreeLS:treePoints]{TreeLS::treePoints()}} with the \code{\link[TreeLS:trp.crop]{TreeLS::trp.crop()}} algorithm
\item Flag only the stem points using \code{\link[TreeLS:stemPoints]{TreeLS::stemPoints()}} with the \code{\link[TreeLS:stm.hough]{TreeLS::stm.hough()}} algorithm
\item Perform DBH estimation using \code{\link[TreeLS:tlsInventory]{TreeLS::tlsInventory()}} with the \code{\link[TreeLS:shapeFit]{TreeLS::shapeFit()}} algorithm
}
}
\examples{
 \dontrun{
 o <- "../data"
 i <- "../data/normlasdata"
 r <- cloud2trees::treels_stem_dbh(folder = i, outfolder = o)
 r \%>\% names()
 }
}
\references{
https://github.com/tiagodc/TreeLS
}
