% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_cbh.R
\name{trees_cbh}
\alias{trees_cbh}
\title{Estimate CBH using tree crown polygons and normalized point cloud data}
\usage{
trees_cbh(
  trees_poly,
  norm_las = NULL,
  tree_sample_n = NA,
  tree_sample_prop = NA,
  which_cbh = "lowest",
  estimate_missing_cbh = TRUE,
  min_vhp_n = 3,
  voxel_grain_size_m = 1,
  dist_btwn_bins_m = 1,
  min_fuel_layer_ht_m = 1,
  lad_pct_gap = 25,
  lad_pct_base = 25,
  num_jump_steps = 1,
  min_lad_pct = 10,
  frst_layer_min_ht_m = 1,
  force_same_crs = F
)
}
\arguments{
\item{trees_poly}{sf. A \code{sf} class object with POLYGON geometry (see \code{\link[sf:st_geometry_type]{sf::st_geometry_type()}}), the program will use the data "as-is" and only require the \code{treeID} and \code{tree_height_m} columns.}

\item{norm_las}{character. a directory with nomalized las files, the path of a single .laz|.las file", -or- an object of class \code{LAS}.
It is your responsibility to ensure that the point cloud is projected the same as the \code{trees_poly} data}

\item{tree_sample_n, tree_sample_prop}{numeric. Provide either \code{tree_sample_n}, the number of trees, or \code{tree_sample_prop}, the
proportion of the trees to attempt to extract a CBH from the point cloud for.
If neither are supplied, \code{tree_sample_n = 155} will be used. If both are supplied, \code{tree_sample_n} will be used.
Increasing \code{tree_sample_prop} toward one (1) will increase the processing time, perhaps significantly depending on the number of trees in the \code{trees_poly} data.}

\item{which_cbh}{character. One of: "lowest"; "highest"; or "max_lad". See Viedma et al. (2024) reference.
\itemize{
\item "lowest" - Height of the CBH of the segmented tree based on the last distance found in its profile
\item "highest" - Height of the CBH of the segmented tree based on the maximum distance found in its profile
\item "max_lad" - Height of the CBH of the segmented tree based on the maximum LAD percentage
}}

\item{estimate_missing_cbh}{logical. even if the \code{tree_sample_prop} parameter is set to "1", it is not likely that CBH will be extracted successfully from every tree.
Should the missing CBH values be estimated using the tree height and location information based on trees for which CBH is successfully extracted?}

\item{min_vhp_n}{numeric. the minimum number of vertical height profiles (VHPs) needed to estimate a CBH.}

\item{voxel_grain_size_m}{numeric. horizontal resolution (suggested 1 meter for lad profiles and 10 meters for LAI maps). See \code{grain.size} in \code{\link[leafR:lad.voxels]{leafR::lad.voxels()}}}

\item{dist_btwn_bins_m}{numeric. value for the actual height bin step (in meters). See \code{step} in \code{\link[LadderFuelsR:get_gaps_fbhs]{LadderFuelsR::get_gaps_fbhs()}}}

\item{min_fuel_layer_ht_m}{numeric. value for the actual minimum base height (in meters). See \code{min_height} in \code{\link[LadderFuelsR:get_gaps_fbhs]{LadderFuelsR::get_gaps_fbhs()}}}

\item{lad_pct_gap}{numeric. value of the percentile threshold used to identify gaps (default percentile 25th). See \code{perc_gap} in \code{\link[LadderFuelsR:get_gaps_fbhs]{LadderFuelsR::get_gaps_fbhs()}}}

\item{lad_pct_base}{numeric. value of the percentile threshold used to identify fuels layers base height (default percentile 25th). See \code{perc_base} in \code{\link[LadderFuelsR:get_gaps_fbhs]{LadderFuelsR::get_gaps_fbhs()}}}

\item{num_jump_steps}{numeric. value for the number of height bin steps that can be jumped to reshape fuels layers. See \code{number_steps} in \code{\link[LadderFuelsR:get_real_fbh]{LadderFuelsR::get_real_fbh()}}}

\item{min_lad_pct}{numeric. value for the minimum required LAD percentage in a fuel layer. See \code{threshold} in \code{\link[LadderFuelsR:get_layers_lad]{LadderFuelsR::get_layers_lad()}}}

\item{frst_layer_min_ht_m}{numeric. value for the depth height of the first fuel layer. If the first fuel layer has the maximum LAD and its depth is greater than the indicated value, then this fuel layer is considered as the CBH of the tree. On the contrary, if its depth is <= the value, the CBH with maximum LAD will be the second fuel layer, although it has not the maximum LAD. See \code{hdepth1_height} in \code{\link[LadderFuelsR:get_cbh_metrics]{LadderFuelsR::get_cbh_metrics()}}}

\item{force_same_crs}{logical. force the same crs between the point cloud and polygon if confident that data are in same projection.
data created by a \code{cloud2trees} pipeline (e.g. \code{\link[=cloud2raster]{cloud2raster()}}) will always have the same projection even if not recognized by \code{lidR} functions}
}
\value{
Returns a spatial data frame of individual trees.
}
\description{
\code{trees_cbh()} uses the input tree crown polygons (e.g. as exported by \code{\link[=raster2trees]{raster2trees()}}) with the columns
\code{treeID} and \code{tree_height_m} to estimate tree CBH using height normalized point cloud data (e.g. as exported by \code{\link[=cloud2raster]{cloud2raster()}}).

CBH is extracted directly from the height normalized point cloud using the process outlined in Viedma et al. (2024) and implemented via \code{\link[=ladderfuelsr_cbh]{ladderfuelsr_cbh()}}.

There are likely to be trees for which there is insufficient data in the point cloud to successfully estimate CBH. The user can elect to estimate missing CBH values which is accomplished via:
\itemize{
\item Attempt to extract CBH from the sample of trees elected by the user (\code{tree_sample_n},\code{tree_sample_prop} parameter) using \code{\link[=ladderfuelsr_cbh]{ladderfuelsr_cbh()}}
\item Successfully extracted CBH trees become training data used to estimate the height-CBH allometry relationship that is spatially informed using the relative tree location compared to the training data
\item The height and location predicting CBH model built from the point cloud training data is used to predict CBH for the non-training (i.e. missing CBH) data
}
}
\examples{
 \dontrun{
 library(tidyverse)
 # example tree crown polygons
 f <- system.file(package = "cloud2trees","extdata","crowns_poly.gpkg")
 crowns <- sf::st_read(f, quiet = T)
 # example normalized las files are in this directory
 norm_d <- system.file(package = "cloud2trees","extdata","norm_las")
 # now run the trees_cbh()
 trees_cbh_ans <- trees_cbh(
    trees_poly = crowns
    , norm_las = norm_d
    , estimate_missing_cbh = T
    , force_same_crs = T
   )
 # what?
 trees_cbh_ans \%>\% class()
 trees_cbh_ans \%>\% dplyr::select(treeID,tidyselect::contains("cbh")) \%>\% dplyr::glimpse()
 # spatial polygons
 trees_cbh_ans \%>\% ggplot2::ggplot() +
   ggplot2::geom_sf(ggplot2::aes(fill=tree_cbh_m,color=is_training_cbh))
 # relationship between height and cbh
 trees_cbh_ans \%>\%
    ggplot2::ggplot(
      ggplot2::aes(x = tree_height_m, y = tree_cbh_m, color=is_training_cbh)
     ) +
    ggplot2::geom_point()
 # tabulate training data
 trees_cbh_ans \%>\%
   sf::st_drop_geometry() \%>\%
   dplyr::count(is_training_cbh)
 }
}
\references{
\itemize{
\item \url{https://doi.org/10.1111/2041-210X.14427}
Viedma, O., Silva, C. A., Moreno, J. M., & Hudak, A. T. (2024). LadderFuelsR: A new automated tool for vertical fuel continuity analysis and crown base height detection using light detection and ranging. Methods in Ecology and Evolution.
\url{https://github.com/olgaviedma/LadderFuelsR}
\item \url{https://doi.org/10.3390/rs11010092}
Almeida, D. R. A. D., Stark, S. C., Shao, G., Schietti, J., Nelson, B. W., Silva, C. A., ... & Brancalion, P. H. S. (2019). Optimizing the remote detection of tropical rainforest structure with airborne lidar: Leaf area profile sensitivity to pulse density and spatial sampling. Remote Sensing, 11(1), 92.
\url{https://github.com/DRAAlmeida/leafR}
}
}
