% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_species.R
\name{trees_species}
\alias{trees_species}
\title{Estimate DBH for a tree list based on height}
\usage{
trees_species(
  tree_list,
  crs = NA,
  study_boundary = NA,
  input_foresttype_dir = paste0(system.file(package = "cloud2trees"),
    "/extdata/foresttype"),
  max_search_dist_m = 1000
)
}
\arguments{
\item{tree_list}{data.frame. A data frame with the columns \code{treeID}, \code{tree_x}, \code{tree_y}, and \code{tree_height_m}.
If an \code{sf} class object with POINT geometry (see \code{\link[sf:st_geometry_type]{sf::st_geometry_type()}}), the program will use the data "as-is" and only require the \code{treeID} and \code{tree_height_m} columns.}

\item{crs}{string. A crs string as returned from \code{\link[sf:st_crs]{sf::st_crs()}} or the EPSG code of the x,y coordinates.
Defaults to the crs of the \code{tree_list} data if of class "sf".}

\item{study_boundary}{sf. The boundary of the study are to define the area of the regional model.
If no boundary given, regional model will be built from location of trees in the tree list.}

\item{input_foresttype_dir}{directory where Forest Type Groups data exists. Use \code{\link[=get_foresttype]{get_foresttype()}} first.}

\item{max_search_dist_m}{number. Maximum search distance (m) to obtain forest type group data for trees in \code{tree_list} that overlap with non-forest data in the original Wilson (2023) data.
Larger search distances will increase processing time and possibly result in memory issues.}
}
\value{
Returns a list of objects: tree_list = spatial data frame of individual trees; foresttype_rast = raster of forest types in the area.
}
\description{
\code{trees_species()} uses the input tree list (e.g. as exported by \code{\link[=raster2trees]{raster2trees()}}) with the columns
\code{treeID}, \code{tree_x}, \code{tree_y} to attach species information using USDA Forest Inventory and Analysis (FIA) codes.
If a spatial data frame of points is the input tree list, then the columns \code{tree_x}, \code{tree_y} are not required.

FIA Forest Type Group Code is attached to each tree in the tree list based on the spatial overlap with the Forest Type Groups of the Continental United States dataset \href{https://www.arcgis.com/home/item.html?id=10760c83b9e44923bd3c18efdaa7319d}{Wilson 2023}.

The simplified process for attaching forest type group to a tree is:
\itemize{
\item Forest type group 30-m raster (Wilson 2023) was aggregated to 90-m to make the data more accessible over the entire continental US
\item Nearest neighbor imputation is used to fill forest type data if a tree falls inside a no-forest cell in the original data
\item The FIA forest type group is applied to a tree based on spatial overlap
}
}
\examples{
 \dontrun{
 # example tree list
 tl <- dplyr::tibble(
     treeID = c(1:21)
     , tree_x = rnorm(n=21, mean = 458064, sd = 11)
     , tree_y = rnorm(n=21, mean = 4450074, sd = 11)
     , tree_height_m = exp(rgamma(n = 21, shape = (7/4)^2, rate = (4^2)/7))
   )
 # call the function
 tl_comp <- trees_competition(tree_list = tl, crs = "32613")
 # what?
 tl_comp \%>\% class()
 tl_comp \%>\% dplyr::select(tidyselect::starts_with("comp_")) \%>\% dplyr::glimpse()
 tl_comp \%>\% ggplot2::ggplot() + ggplot2::geom_sf(ggplot2::aes(color=comp_dist_to_nearest_m))
 }
}
\references{
\itemize{
\item \href{https://www.arcgis.com/home/item.html?id=10760c83b9e44923bd3c18efdaa7319d}{Forest Type Groups of the Continental United States}
Wilson, B.T. (2023). Forest Type Groups of the Continental United States.
}
}
