% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rast_points.R
\name{crop_raster_match_points}
\alias{crop_raster_match_points}
\title{internal functions to extract raster values at point locations}
\usage{
crop_raster_match_points(
  points,
  rast,
  study_boundary = NA,
  max_search_dist_m = 1000
)
}
\arguments{
\item{points}{sf.}

\item{rast}{SpatRaster.}

\item{study_boundary}{sf.}

\item{max_search_dist_m}{numeric.}
}
\description{
internal functions to extract raster values at point locations used by \code{\link[=trees_type]{trees_type()}} and tree_biomass() for LANDFIRE rasters
for the most part, these functions take point and raster data as inputs
the general process is (see \code{\link[=trees_type]{trees_type()}} for example):
\itemize{
\item crop_raster_match_points()
\item check for undesirable or NA values in the point_values from crop_raster_match_points()
\item if undesirable values:
\itemize{
\item reclass_landfire_rast() or reclass_foresttype_rast()
\item agg_fill_rast_match_points(), see details above function definition
\item use the point_values from agg_fill_rast_match_points() if not null
\item otherwise, use the point_values from crop_raster_match_points()
}
}
}
\keyword{internal}
