% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_foresttype.R
\name{get_foresttype}
\alias{get_foresttype}
\title{Download Forest Type Groups of the Continental United States data}
\usage{
get_foresttype(savedir = NULL, force = F)
}
\arguments{
\item{savedir}{Optional directory to save data in a new location. Defaults to package contents.}

\item{force}{Whether to overwrite existing data}
}
\description{
The Forest Type Groups of the Continental United States data is used to estimate individual tree forest type group.
See \code{\link[=trees_species]{trees_species()}}
}
\examples{
 \dontrun{
 get_foresttype()
 }
}
\references{
\itemize{
\item \href{https://www.arcgis.com/home/item.html?id=10760c83b9e44923bd3c18efdaa7319d}{Forest Type Groups of the Continental United States}
Wilson, B.T. (2023). Forest Type Groups of the Continental United States.
}
}
