% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasr_pipeline.R
\name{lasr_pipeline}
\alias{lasr_pipeline}
\title{Create lasR pipeline to process a las grid tile created via \code{chunk_las_catalog}}
\usage{
lasr_pipeline(
  processing_grid_num = 1,
  process_data,
  keep_intrmdt = F,
  dtm_res_m = 1,
  chm_res_m = 0.25,
  min_height = 2,
  max_height = 70,
  dtm_dir = getwd(),
  chm_dir = getwd(),
  classify_dir = getwd(),
  normalize_dir = getwd()
)
}
\arguments{
\item{processing_grid_num}{numeric. processing_grid column in the data.frame created via \code{chunk_las_catalog}}

\item{process_data}{data.frame. data.frame created via \code{chunk_las_catalog}}

\item{keep_intrmdt}{logical. this process writes intermediate data to the disk, keep those intermediate files (classfied, normalized, stem las files)?}

\item{dtm_res_m}{numeric. The desired resolution of the DTM produced in meters.}

\item{chm_res_m}{numeric. The desired resolution of the CHM produced in meters.}

\item{min_height}{numeric. Set the minimum height (m) for individual tree detection}

\item{max_height}{numeric. Set the maximum height (m) for the canopy height model}

\item{dtm_dir}{string. The path of a folder to write the tiled DTM files to.}

\item{chm_dir}{string. The path of a folder to write the tiled CHM files to.}

\item{classify_dir}{string. The path of a folder to write the classified .las files to.}

\item{normalize_dir}{string. The path of a folder to write the normalized .las files to.}
}
\value{
A \code{lasR} pipeline answer list
}
\description{
Create lasR pipeline to process a las grid tile created via \code{chunk_las_catalog}
}
\references{
https://r-lidar.github.io/lasR/index.html
}
\keyword{internal}
