% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_biomass.R
\name{trees_biomass}
\alias{trees_biomass}
\title{Estimate tree biomass (or crown biomass) for a tree list}
\usage{
trees_biomass(
  tree_list,
  crs = NA,
  study_boundary = NA,
  input_landfire_dir = NULL,
  input_foresttype_dir = NULL,
  method = NA
)
}
\arguments{
\item{tree_list}{data.frame. A data frame with the columns \code{treeID}, \code{tree_x}, and \code{tree_y}.
If an \code{sf} class object with POINT geometry (see \code{\link[sf:st_geometry_type]{sf::st_geometry_type()}}), the program will use the data "as-is" and only require the \code{treeID} column.
Other required columns include:
\itemize{
\item \code{crown_area_m2}, \code{tree_height_m} (e.g. as exported by \code{\link[=raster2trees]{raster2trees()}})
\item \code{tree_cbh_m} (e.g. as exported by \code{\link[=trees_cbh]{trees_cbh()}})
\item and one of \code{dbh_cm}, \code{dbh_m}, or  \code{basal_area_m2} (e.g. as exported by \code{\link[=trees_dbh]{trees_dbh()}})
}}

\item{crs}{string. A crs string as returned from \code{\link[sf:st_crs]{sf::st_crs()}} or the EPSG code of the x,y coordinates.
Defaults to the crs of the \code{tree_list} data if of class "sf".}

\item{study_boundary}{sf. The boundary of the study are to define the area of interest which may extend beyond the space with trees.
If no boundary given, the AOI will be built from location of trees in the tree list.}

\item{input_landfire_dir}{directory where LANDFIRE CBD data exists. Use \code{\link[=get_landfire]{get_landfire()}} first.}

\item{input_foresttype_dir}{directory where Forest Type Groups data exists. Use \code{\link[=get_foresttype]{get_foresttype()}} first.}

\item{method}{character. one (e.g. \code{"landfire"}) or multiple (e.g. \code{c("cruz","landfire")}) of the following biomass estimation methods:
\itemize{
\item Tree crown biomass in kilograms:
\itemize{
\item "landfire" - based on LANDFIRE's Forest Canopy Bulk Density (CBD) data (\code{\link[=trees_biomass_landfire]{trees_biomass_landfire()}})
\item "cruz" - based on Cruz et al. (2003) canopy fuel stratum equations (\code{\link[=trees_biomass_cruz]{trees_biomass_cruz()}})
}
\item Tree total above ground biomass in kilograms:
\itemize{
\item coming soon
}
}}
}
\value{
Returns a list of objects: tree_list = spatial data frame of individual trees
; stand_cell_data_landfire = data frame of stands/cells in same projection as the LANDFIRE raster data
; stand_cell_data_cruz = data frame of stands/cells in same projection as the FIA forest type group raster data

See code in examples.
}
\description{
\code{trees_biomass()} streamlines the process for estimating individual tree biomass in kilograms, or the component biomass of the tree crown in kilograms.
Users can select one or all of the following methods available in the package for estimating biomass:
\itemize{
\item Tree crown biomass in kilograms:
\itemize{
\item "landfire" - based on LANDFIRE's Forest Canopy Bulk Density (CBD) data (\code{\link[=trees_biomass_landfire]{trees_biomass_landfire()}})
\item "cruz" - based on Cruz et al. (2003) canopy fuel stratum equations (\code{\link[=trees_biomass_cruz]{trees_biomass_cruz()}})
}
\item Tree total above ground biomass in kilograms:
\itemize{
\item coming soon
}
}

If multiple methods are selected (e.g. \code{method = c("cruz","landfire")}), then the program will compile the
biomass estimates and return one tree list.
}
\examples{
 \dontrun{
library(tidyverse)
library(sf)
# use the tree list that ships with the package
f <- system.file(package = "cloud2trees", "extdata", "crowns_poly.gpkg")
tl <- sf::st_read(f)
tl \%>\% dplyr::glimpse()
# call trees_biomass and get multiple biomass estimates
trees_biomass_ans <- trees_biomass(tree_list = tl, method = c("landfire","cruz"))
# what did we get back?
trees_biomass_ans \%>\% names()
# check out the tree list
trees_biomass_ans$tree_list \%>\% dplyr::glimpse()
# check out the landfire stand data
trees_biomass_ans$stand_cell_data_landfire \%>\% dplyr::filter(trees>0) \%>\% dplyr::glimpse()
# plot tree landfire crown biomass estimate
trees_biomass_ans$tree_list \%>\%
  ggplot2::ggplot(
    mapping = ggplot2::aes(
      x = tree_height_m
      , y = landfire_crown_biomass_kg
      , color = crown_area_m2
    )
  ) +
  ggplot2::geom_point()
# plot tree cruz crown biomass estimate
trees_biomass_ans$tree_list \%>\%
  ggplot2::ggplot(
    mapping = ggplot2::aes(
      x = tree_height_m
      , y = cruz_crown_biomass_kg
      , color = crown_area_m2
    )
  ) +
  ggplot2::geom_point()
# plot tree landfire vs. cruz crown biomass estimate
trees_biomass_ans$tree_list \%>\%
  ggplot2::ggplot(
    mapping = ggplot2::aes(
      x = landfire_crown_biomass_kg, y = cruz_crown_biomass_kg
    )
  ) +
  ggplot2::geom_abline(lwd = 1.5) +
  ggplot2::geom_smooth(method = "lm", se=F, color = "gray", linetype = "dashed") +
  ggplot2::geom_point(ggplot2::aes(color = tree_height_m)) +
  ggplot2::scale_x_continuous(
    limits = c(0
      , max(trees_biomass_ans$tree_list$cruz_crown_biomass_kg)
    )
  ) +
  ggplot2::scale_y_continuous(
    limits = c(0
      , max(trees_biomass_ans$tree_list$cruz_crown_biomass_kg)
    )
  )
# get the projection for the stand cell data
epsg_code <- trees_biomass_ans$stand_cell_data_landfire$rast_epsg_code[1] \%>\% as.numeric()
# plot the stand cell data with trees overlaid
trees_biomass_ans$stand_cell_data_landfire \%>\%
  dplyr::filter(trees>0) \%>\%
  ggplot2::ggplot() +
  ggplot2::geom_tile(ggplot2::aes(x=x,y=y,fill = landfire_stand_kg_per_m3), color = "gray44") +
  ggplot2::geom_text(ggplot2::aes(x=x,y=y,label = trees), color = "white") +
  ggplot2::geom_sf(
    data = trees_biomass_ans$tree_list \%>\% sf::st_transform(crs = epsg_code)
    , ggplot2::aes(color = cruz_crown_biomass_kg)
  ) +
  ggplot2::labs(
    fill="stand kg/m3", color = "landfire\ncrown kg"
    , caption = "# trees shown in cell"
  ) +
  ggplot2::scale_fill_viridis_c(option = "rocket", na.value = "gray", direction = -1) +
  ggplot2::scale_color_viridis_c(option = "viridis", na.value = "gray22", begin = 0.6) +
  ggplot2::theme_void()
 }
}
\references{
See references in:
\itemize{
\item \code{\link[=trees_biomass_landfire]{trees_biomass_landfire()}}
\item \code{\link[=trees_biomass_cruz]{trees_biomass_cruz()}}
}
}
