% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_biomass_landfire.R
\name{trees_biomass_landfire}
\alias{trees_biomass_landfire}
\title{Estimate tree biomass for a tree list using LANDFIRE data}
\usage{
trees_biomass_landfire(
  tree_list,
  crs = NA,
  study_boundary = NA,
  input_landfire_dir = NULL,
  max_search_dist_m = 1000
)
}
\arguments{
\item{tree_list}{data.frame. A data frame with the columns \code{treeID}, \code{tree_x}, \code{tree_y}, and \code{tree_height_m}.
If an \code{sf} class object with POINT geometry (see \code{\link[sf:st_geometry_type]{sf::st_geometry_type()}}), the program will use the data "as-is" and only require the \code{treeID} column.}

\item{crs}{string. A crs string as returned from \code{\link[sf:st_crs]{sf::st_crs()}} or the EPSG code of the x,y coordinates.
Defaults to the crs of the \code{tree_list} data if of class "sf".}

\item{study_boundary}{sf. The boundary of the study are to define the area of the regional model.
If no boundary given, regional model will be built from location of trees in the tree list.}

\item{input_landfire_dir}{directory where LANDFIRE Forest Canopy Bulk Density data exists. Use \code{\link[=get_landfire]{get_landfire()}} first.}

\item{max_search_dist_m}{number. Maximum search distance (m) to obtain forest type group data for trees in \code{tree_list} that overlap with non-forest data in the original LANDFIRE data.
Larger search distances will increase processing time and possibly result in memory issues.}
}
\value{
Returns a list of objects: tree_list = spatial data frame of individual trees; landfire_rast = raster of forest types in the area.
}
\description{
\code{trees_biomass_landfire()} uses the input tree list (e.g. as exported by \code{\link[=raster2trees]{raster2trees()}}) with the columns
\code{treeID}, \code{tree_x}, \code{tree_y} to attach an estimate of tree crown biomass using LANDFIRE's Forest Canopy Bulk Density (CBD) data
produced jointly by the U.S. Department of Agriculture and U.S. Department of the Interior.
If a spatial data frame of points is the input tree list, then the columns \code{tree_x}, \code{tree_y} are not required.

LANDFIRE's Forest Canopy Bulk Density (CBD) data is attached to each tree in the tree list based on the spatial overlap with the raster dataset (see refernces).
Canopy Bulk Density is mass of flammable material per unit volume of the tree crown typically expressed in units of mass per unit volume (e.g., kilograms per cubic meter).

The simplified process for attaching tree crown biomass in kilograms to a tree is:
\itemize{
\item Nearest neighbor imputation is used to fill LANDFIRE data if a tree falls inside a non-forest cell in the original data
\item The LANDFIRE estimate of CBD is distributed across the individual trees that fall in a raster cell by:
\enumerate{
\item asdf
\item derp
\item hey
}
}
}
\examples{
 \dontrun{
 # example tree list
 tl <- dplyr::tibble(
     treeID = c(1:21)
     , tree_x = rnorm(n=21, mean = 458064, sd = 11)
     , tree_y = rnorm(n=21, mean = 4450074, sd = 11)
   )
 # call the function
 tl_type <- trees_type(tree_list = tl, crs = "32613")
 # what?
 tl_type \%>\% class()
 # a list, but what is in it?
 tl_type \%>\% names()
 # plot the tree_list spatial points
 tl_type$tree_list \%>\% ggplot2::ggplot() + ggplot2::geom_sf(ggplot2::aes(color=forest_type_group))
 # plot the landfire_rast raster
 tl_type$landfire_rast \%>\% terra::plot()
 }
}
\references{
\itemize{
\item \href{https://landfire.gov/fuel/cbd}{LANDFIRE Forest Canopy Bulk Density (CBD)}
U.S. Department of Agriculture and U.S. Department of the Interior.
}
}
