% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_biomass_landfire.R
\name{trees_biomass_landfire}
\alias{trees_biomass_landfire}
\title{Estimate tree crown biomass for a tree list using LANDFIRE data}
\usage{
trees_biomass_landfire(
  tree_list,
  crs = NA,
  study_boundary = NA,
  input_landfire_dir = NULL
)
}
\arguments{
\item{tree_list}{data.frame. A data frame with the columns \code{treeID}, \code{tree_x}, and \code{tree_y}.
If an \code{sf} class object with POINT geometry (see \code{\link[sf:st_geometry_type]{sf::st_geometry_type()}}), the program will use the data "as-is" and only require the \code{treeID} column.
Other required columns include:
\itemize{
\item \code{crown_area_m2}, \code{tree_height_m} (e.g. as exported by \code{\link[=raster2trees]{raster2trees()}})
\item \code{tree_cbh_m} (e.g. as exported by \code{\link[=trees_cbh]{trees_cbh()}})
\item and one of \code{dbh_cm}, \code{dbh_m}, or  \code{basal_area_m2} (e.g. as exported by \code{\link[=trees_dbh]{trees_dbh()}})
}}

\item{crs}{string. A crs string as returned from \code{\link[sf:st_crs]{sf::st_crs()}} or the EPSG code of the x,y coordinates.
Defaults to the crs of the \code{tree_list} data if of class "sf".}

\item{study_boundary}{sf. The boundary of the study are to define the area of interest which may extend beyond the space with trees.
If no boundary given, the AOI will be built from location of trees in the tree list.}

\item{input_landfire_dir}{directory where LANDFIRE CBD data exists. Use \code{\link[=get_landfire]{get_landfire()}} first.}
}
\value{
Returns a list of objects: tree_list = spatial data frame of individual trees
; stand_cell_data = data frame of stands/cells in same projection as the LANDFIRE raster data

See code in examples.
}
\description{
\code{trees_biomass_landfire()} uses the input tree list (e.g. as exported by \code{\link[=raster2trees]{raster2trees()}}) with the columns
\code{treeID}, \code{tree_x}, \code{tree_y} to attach an estimate of tree crown biomass using LANDFIRE's Forest Canopy Bulk Density (CBD) data
produced jointly by the U.S. Department of Agriculture and U.S. Department of the Interior.

If a spatial data frame of points is the input tree list, then the columns \code{tree_x}, \code{tree_y} are not required.
Other required columns include:
\itemize{
\item \code{crown_area_m2}, \code{tree_height_m} (e.g. as exported by \code{\link[=raster2trees]{raster2trees()}})
\item \code{tree_cbh_m} (e.g. as exported by \code{\link[=trees_cbh]{trees_cbh()}})
\item and one of \code{dbh_cm}, \code{dbh_m}, or  \code{basal_area_m2} (e.g. as exported by \code{\link[=trees_dbh]{trees_dbh()}})
}

LANDFIRE's Forest Canopy Bulk Density (CBD) data is attached to each tree in the tree list based
on the spatial overlap with the raster data set (see references).
Canopy Bulk Density is mass of flammable material per unit volume of the tree
crown typically expressed in units of mass per unit volume (e.g., kilograms per cubic meter).

The process for estimating tree crown biomass in kilograms is:
\itemize{
\item Nearest neighbor imputation is used to fill LANDFIRE data if a tree falls inside a non-forest cell in the original data
\item The LANDFIRE estimate of CBD is distributed across the individual trees that fall in a raster cell by:
\enumerate{
\item At the stand level (i.e. raster cell), aggregate the tree level data within the stand to obtain:
}
\itemize{
\item \code{mean_crown_length_m = mean(crown_length_m)}, where tree \code{crown_length_m = tree_height_m - tree_cbh_m}
\item \code{sum_crown_volume_m3 = sum(crown_volume_m3)}, where tree \code{crown_volume_m3 = (4/3) * pi * ((crown_length_m/2)) * ((crown_dia_m/2)^2)}
}
\enumerate{
\item At the stand level (i.e. raster cell), determine the area of the stand that overlaps (\code{overlap_area_m2}) with the AOI defined as the \code{study_boundary} parameter (see below) or the bounding box of all the trees
\item At the stand level (i.e. raster cell), get the LANDFIRE estimate of CBD in kilograms per cubic meter (\code{landfire_stand_kg_per_m3})
\item At the stand level (i.e. raster cell), get canopy fuel loading (CFL) in kilograms per cubic meter (\code{kg_per_m2 = mean_crown_length_m * landfire_stand_kg_per_m3})
\item At the stand level (i.e. raster cell), get the stand biomass in kilograms (\code{biomass_kg = kg_per_m2 * overlap_area_m2})
\item At the stand level (i.e. raster cell), the single tree CBD in kilograms per cubic meter will be a constant (\code{landfire_tree_kg_per_m3 = biomass_kg / sum_crown_volume_m3})
\item Attach the the single tree CBD in kilograms per cubic meter to the tree level based on raster cell spatial overlap
\item Calculate individual tree crown mass in kilograms as \code{landfire_crown_biomass_kg = landfire_tree_kg_per_m3 * crown_volume_m3}
}
}
}
\examples{
 \dontrun{
library(tidyverse)
library(sf)
my_n <- 122
# fake tree list
tl <- dplyr::tibble(
    treeID = c(1:my_n)
    , tree_x = rnorm(n=my_n, mean = 458064, sd = 33)
    , tree_y = rnorm(n=my_n, mean = 4450074, sd = 33)
    , tree_height_m = rnorm(n=my_n, mean = 10, sd = 7)
  ) \%>\%
  dplyr::mutate(
    tree_height_m = ifelse(tree_height_m<1.37, 1.37, tree_height_m) # above DBH
    , crown_area_m2 = 0.47+(0.49*tree_height_m)
    , tree_cbh_m = 0.72+(0.53*tree_height_m)
    , dbh_cm = -2.3+(2.14*tree_height_m)
  )
# how does our fake tree list look?
tl \%>\% dplyr::glimpse()
tl \%>\% ggplot2::ggplot() + ggplot2::geom_point(ggplot2::aes(x = dbh_cm, y = tree_height_m))
# call the function
tl_landfire <- trees_biomass_landfire(tree_list = tl, crs = "32613")
# what is in it?
tl_landfire \%>\% names()
# look at the trees
tl_landfire$tree_list \%>\% dplyr::glimpse()
# look at the stand
tl_landfire$stand_cell_data \%>\% dplyr::filter(!is.na(trees)) \%>\% dplyr::glimpse()
# the stand CBD
 tl_landfire$stand_cell_data \%>\%
   dplyr::filter(trees>0) \%>\%
   sf::st_drop_geometry() \%>\%
   dplyr::count(landfire_stand_kg_per_m3)
# get the projection for the stand cell data
epsg_code <- tl_landfire$stand_cell_data$rast_epsg_code[1] \%>\% as.numeric()
# plot the stand cell data with trees overlaid
tl_landfire$stand_cell_data \%>\%
  ggplot2::ggplot() +
  ggplot2::geom_tile(ggplot2::aes(x=x,y=y,fill = landfire_stand_kg_per_m3), color = "gray44") +
  ggplot2::geom_text(ggplot2::aes(x=x,y=y,label = trees), color = "white") +
  ggplot2::geom_sf(
    data = tl_landfire$tree_list \%>\% sf::st_transform(crs = epsg_code)
    , ggplot2::aes(color = landfire_crown_biomass_kg)
  ) +
  ggplot2::labs(fill="stand kg/m3", color = "tree crown kg", caption = "# trees shown in cell") +
  ggplot2::scale_fill_viridis_c(option = "rocket", na.value = "gray", direction = -1) +
  ggplot2::scale_color_viridis_c(option = "viridis", na.value = "gray22", begin = 0.6) +
  ggplot2::theme_void()
 }
}
\references{
\itemize{
\item \href{https://landfire.gov/fuel/cbd}{LANDFIRE Forest Canopy Bulk Density (CBD)}
U.S. Department of Agriculture and U.S. Department of the Interior.
}
}
